+++ lib/message.ts
@@ -105,12 +105,15 @@ export class Message {
 
   toJson() {
     debug('toJson')
+    const quotedMessage = this.getQuotedMessage()
     return {
       chatId: this.getChatId(),
       duration: this.getDuration(),
       file: this.getFile(),
       fromId: this.getFromId(),
       id: this.getId(),
+      quotedText: this.getQuotedText(),
+      quotedMessageId: quotedMessage ? quotedMessage.getId() : null,
       receivedTimestamp: this.getReceivedTimestamp(),
       sortTimestamp: this.getSortTimestamp(),
       text: this.getText(),
@@ -167,6 +170,15 @@ export class Message {
 
   getId(): number {
     return binding.dcn_msg_get_id(this.dc_msg)
+  }
+
+  getQuotedText(): string{
+    return binding.dcn_msg_get_quoted_text(this.dc_msg)
+  }
+
+  getQuotedMessage(): Message|null{
+    const dc_msg = binding.dcn_msg_get_quoted_msg(this.dc_msg)
+    return dc_msg ? new Message(dc_msg): null
   }
 
   getReceivedTimestamp(): number {
@@ -278,6 +290,10 @@ export class Message {
   setLocation(longitude: number, latitude: number) {
     binding.dcn_msg_set_location(this.dc_msg, longitude, latitude)
     return this
+  }
+
+  setQuote(qoutedMessage: Message) {
+    binding.dcn_msg_set_quote(this.dc_msg, qoutedMessage)
   }
 
   setText(text: string) {